+++ .github/workflows/build-enterprise-warehouse-backend.yml
@@ -1,6 +1,6 @@
 # This is a basic workflow to help you get started with Actions
 
-name: build-message-sink
+name: build-enterprise-warehouse-backend
 
 # Controls when the action will run. 
 on:
@@ -9,10 +9,10 @@ on:
 
 defaults:
   run:
-    working-directory: src/Message.Sink
+    working-directory: src/EnterpriseWarehouse.Backend
 
 env:
-  IMAGE_NAME: message-sink
+  IMAGE_NAME: enterprise-warehouse-backend
 
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
