+++ src/Contonance.Backend/appsettings.json
@@ -1,0 +1,17 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "ApplicationInsights": {
+    "LogLevel": {
+      "Default": "Information"
+    }
+  },
+  "HttpClient": {
+    "EnableRetry": false,
+    "EnableBreaker": false
+  }
+}
