+++ src/EnterpriseWarehouse.Backend/CloudRoleNameInitializer.cs
