+++ src/EnterpriseWarehouse.Backend/DeviceMessage.cs
