+++ src/EnterpriseWarehouse.Backend/IMessageStorageService.cs
