+++ src/EnterpriseWarehouse.Backend/MessageCosmosTableStorageService.cs
