--- src/Message.Receiver/appsettings.json
@@ -1,17 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ApplicationInsights": {
-    "LogLevel": {
-      "Default": "Information"
-    }
-  },
-  "HttpClient": {
-    "EnableRetry": false,
-    "EnableBreaker": false
-  }
-}
