--- src/Message.Sink/appsettings.json
@@ -1,28 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ApplicationInsights": {
-    "LogLevel": {
-      "Default": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "IpRateLimiting": {
-    "EnableEndpointRateLimiting": true,
-    "StackBlockedRequests": false,
-    "RealIPHeader": "X-Real-IP",
-    "ClientIdHeader": "X-ClientId",
-    "HttpStatusCode": 429,
-    "GeneralRules": [
-      {
-        "Endpoint": "*:/api/*",
-        "Period": "1s",
-        "Limit": 5
-      }
-    ]
-  }
-}
