+++ template.env
@@ -1,9 +1,9 @@
 ASPNETCORE_ENVIRONMENT="Development"
-RECEIVER_URL="http://localhost:5025/api/message/receive"
-SINK_URL="http://localhost:5027/"
+CONTONANCE_BACKEND_URL="http://localhost:5025/api/message/receive"
+ENTERPRISE_WAREHOUSE_BACKEND_URL="http://localhost:5027/"
 Message.Creator__EnableRetry=false
 Message.Creator__EnableBreaker=false
-Message.Receiver__EnableRetry=false
-Message.Receiver__EnableBreaker=false
+Contonance.Backend__EnableRetry=false
+Contonance.Backend__EnableBreaker=false
 IpRateLimiting__EnableEndpointRateLimiting=true
 VERSION="dev version"
